rename authz files and move functions
