+++ controllers/dbaasauthz_controller.go
@@ -19,7 +19,6 @@ package controllers
 import (
 	"context"
 	"reflect"
-	"strings"
 
 	"github.com/RHEcosystemAppEng/dbaas-operator/api/v1alpha1"
 	oauthzv1 "github.com/openshift/api/authorization/v1"
@@ -100,9 +99,7 @@ func (r *DBaaSAuthzReconciler) SetupWithManager(mgr ctrl.Manager) error {
 			&source.Kind{Type: &v1alpha1.DBaaSTenant{}},
 			handler.EnqueueRequestsFromMapFunc(func(o client.Object) []reconcile.Request {
 				namespace := o.(*v1alpha1.DBaaSTenant).Spec.InventoryNamespace
-				return []reconcile.Request{
-					getRequest(namespace),
-				}
+				return getRequest(namespace)
 			}),
 		).
 		WithOptions(
@@ -122,24 +119,6 @@ func (r *DBaaSAuthzReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	}
 
 	return nil
-}
-
-// changes name to match an object's namespace. ensures a namespace is only queued once,
-// instead of once for each object in a namespace
-var nsMapFunc = func(o client.Object) []reconcile.Request {
-	namespace := o.GetNamespace()
-	return []reconcile.Request{
-		getRequest(namespace),
-	}
-}
-
-func getRequest(nameNS string) reconcile.Request {
-	return reconcile.Request{
-		NamespacedName: types.NamespacedName{
-			Name:      nameNS,
-			Namespace: nameNS,
-		},
-	}
 }
 
 // Reconcile all tenant and inventory RBAC
@@ -404,220 +383,4 @@ func (r *DBaaSAuthzReconciler) createRbacObj(newObj, getObj, owner client.Object
 		logger.V(1).Info(kind+" contains edit or list verbs, will not create", name, namespace)
 	}
 	return false, nil
-}
-
-// gets rbac objects for an inventory's users
-func inventoryRbacObjs(inventory v1alpha1.DBaaSInventory, tenantList v1alpha1.DBaaSTenantList) (rbacv1.Role, rbacv1.RoleBinding) {
-	role := rbacv1.Role{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "dbaas-" + inventory.Name + "-inventory-viewer",
-			Namespace: inventory.Namespace,
-		},
-		Rules: []rbacv1.PolicyRule{
-			{
-				APIGroups:     []string{v1alpha1.GroupVersion.Group},
-				Resources:     []string{"dbaasinventories", "dbaasinventories/status"},
-				ResourceNames: []string{inventory.Name},
-				Verbs:         []string{"get"},
-			},
-		},
-	}
-	role.SetGroupVersionKind(rbacv1.SchemeGroupVersion.WithKind("Role"))
-
-	roleBinding := rbacv1.RoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      role.Name + "s",
-			Namespace: role.Namespace,
-		},
-		RoleRef: rbacv1.RoleRef{
-			APIGroup: rbacv1.SchemeGroupVersion.Group,
-			Kind:     "Role",
-			Name:     role.Name,
-		},
-	}
-	roleBinding.SetGroupVersionKind(rbacv1.SchemeGroupVersion.WithKind("RoleBinding"))
-
-	// if inventory.Spec.Authz is nil, use tenant defaults for view access to the Inventory object
-	var users, groups []string
-	if inventory.Spec.Authz.Users == nil && inventory.Spec.Authz.Groups == nil {
-		for _, tenant := range tenantList.Items {
-			if tenant.Spec.InventoryNamespace == inventory.Namespace {
-				users = append(users, tenant.Spec.Authz.Users...)
-				groups = append(groups, tenant.Spec.Authz.Groups...)
-			}
-		}
-	} else {
-		users = inventory.Spec.Authz.Users
-		groups = inventory.Spec.Authz.Groups
-	}
-	users = uniqueStrSlice(users)
-	groups = uniqueStrSlice(groups)
-
-	roleBinding.Subjects = getSubjects(users, groups, role.Namespace)
-
-	return role, roleBinding
-}
-
-// gets rbac objects for a tenant's users
-func tenantRbacObjs(tenant v1alpha1.DBaaSTenant, serviceAdminAuthz, developerAuthz, tenantListAuthz v1alpha1.DBaasUsersGroups) (rbacv1.ClusterRole, rbacv1.ClusterRoleBinding) {
-	clusterRole := rbacv1.ClusterRole{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "dbaas-" + tenant.Name + "-tenant-viewer",
-		},
-		Rules: []rbacv1.PolicyRule{
-			{
-				APIGroups:     []string{v1alpha1.GroupVersion.Group},
-				Resources:     []string{"dbaastenants", "dbaastenants/status"},
-				ResourceNames: []string{tenant.Name},
-				Verbs:         []string{"get"},
-			},
-		},
-	}
-	clusterRole.SetGroupVersionKind(rbacv1.SchemeGroupVersion.WithKind("ClusterRole"))
-
-	clusterRoleBinding := rbacv1.ClusterRoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: clusterRole.Name + "s",
-		},
-		RoleRef: rbacv1.RoleRef{
-			APIGroup: rbacv1.SchemeGroupVersion.Group,
-			Kind:     "ClusterRole",
-			Name:     clusterRole.Name,
-		},
-	}
-	clusterRoleBinding.SetGroupVersionKind(rbacv1.SchemeGroupVersion.WithKind("ClusterRoleBinding"))
-
-	// give view access to all inventory devs and tenant service admins for the Tenant object
-	users := uniqueStrSlice(append(serviceAdminAuthz.Users, developerAuthz.Users...))
-	groups := uniqueStrSlice(append(serviceAdminAuthz.Groups, developerAuthz.Groups...))
-	// remove results of tenant access review, as they don't need the addtl the perms
-	users = removeFromSlice(tenantListAuthz.Users, users)
-	groups = removeFromSlice(tenantListAuthz.Groups, groups)
-
-	clusterRoleBinding.Subjects = getSubjects(users, groups, "")
-
-	return clusterRole, clusterRoleBinding
-}
-
-// verify no edit or list permissions are assigned to a role
-func hasNoEditOrListVerbs(roleObj client.Object) bool {
-	var verbs []string
-	var roleRules []rbacv1.PolicyRule
-	editVerbs := []string{"create", "patch", "update", "delete", "list"}
-
-	kind := roleObj.GetObjectKind().GroupVersionKind().Kind
-	if kind == "Role" {
-		roleRules = roleObj.(*rbacv1.Role).Rules
-	}
-	if kind == "ClusterRole" {
-		roleRules = roleObj.(*rbacv1.ClusterRole).Rules
-	}
-	for _, rules := range roleRules {
-		verbs = append(verbs, rules.Verbs...)
-	}
-	for _, verb := range editVerbs {
-		if contains(verbs, verb) {
-			return false
-		}
-	}
-	return true
-}
-
-// get cumulative authz from all inventories in namespace
-func getDevAuthzFromInventoryList(inventoryList v1alpha1.DBaaSInventoryList, tenant v1alpha1.DBaaSTenant) (developerAuthz v1alpha1.DBaasUsersGroups) {
-	var tenantDefaults bool
-	for _, inventory := range inventoryList.Items {
-		if inventory.Namespace == tenant.Spec.InventoryNamespace {
-			// if inventory.spec.authz is nil, apply authz from tenant.spec.authz.developer as a default
-			if inventory.Spec.Authz.Users == nil && inventory.Spec.Authz.Groups == nil {
-				tenantDefaults = true
-			} else {
-				developerAuthz.Users = append(developerAuthz.Users, inventory.Spec.Authz.Users...)
-				developerAuthz.Groups = append(developerAuthz.Groups, inventory.Spec.Authz.Groups...)
-			}
-		}
-	}
-	if tenantDefaults {
-		developerAuthz.Users = append(developerAuthz.Users, tenant.Spec.Authz.Users...)
-		developerAuthz.Groups = append(developerAuthz.Groups, tenant.Spec.Authz.Groups...)
-	}
-	return developerAuthz
-}
-
-// create a slice of rbac subjects for use in role bindings
-func getSubjects(users, groups []string, namespace string) []rbacv1.Subject {
-	subjects := []rbacv1.Subject{}
-	for _, user := range users {
-		if strings.HasPrefix(user, "system:serviceaccount:") {
-			sa := strings.Split(user, ":")
-			if len(sa) > 3 {
-				subjects = append(subjects, getSubject(sa[3], sa[2], "ServiceAccount"))
-			}
-		} else {
-			subjects = append(subjects, getSubject(user, namespace, "User"))
-		}
-	}
-	for _, group := range groups {
-		subjects = append(subjects, getSubject(group, namespace, "Group"))
-	}
-	if len(subjects) == 0 {
-		return nil
-	}
-
-	return subjects
-}
-
-// create an rbac subject for use in role bindings
-func getSubject(name, namespace, rbacObjectKind string) rbacv1.Subject {
-	subject := rbacv1.Subject{
-		Kind:      rbacObjectKind,
-		Name:      name,
-		Namespace: namespace,
-	}
-	if subject.Kind != "ServiceAccount" {
-		subject.APIGroup = rbacv1.SchemeGroupVersion.Group
-	}
-	return subject
-}
-
-// returns a unique matching subset of the provided slices
-func matchSlices(input1, input2 []string) []string {
-	m := []string{}
-	for _, val1 := range input1 {
-		for _, val2 := range input2 {
-			if val1 == val2 {
-				m = append(m, val1)
-			}
-		}
-	}
-
-	return uniqueStrSlice(m)
-}
-
-// returns a unique subset of the provided slice
-func uniqueStrSlice(input []string) []string {
-	u := make([]string, 0, len(input))
-	m := make(map[string]bool)
-
-	for _, val := range input {
-		if _, ok := m[val]; !ok {
-			m[val] = true
-			u = append(u, val)
-		}
-	}
-
-	return u
-}
-
-// returns a subset of the provided slices, after removing certain entries
-func removeFromSlice(entriesToRemove, fromSlice []string) []string {
-	r := []string{}
-	for _, val := range fromSlice {
-		if !contains(entriesToRemove, val) {
-			r = append(r, val)
-		}
-
-	}
-
-	return r
 }
